r2048: Implemented item 27 of the ROADMAP, removed drbd_md_write() and replaced this with drbd_md_sync(). I guess this was Lars' idea when he created the MD_DIRTY flag.
In case something is modified that is should be writeen to the meta-data
super block later, we mark this by calling drbd_md_mark_dirty(). 

A timer ensures that latest one second later it is synced to this. The
timer is intended as debugging aid, to find the code pathes where the call
to drbd_md_sync() is missing.