r2052: Rewrote drbdadm's scanner/lexer and parser. The previous scanner/lexer was more a parser by itself, and the parser was in its own way obscure as well.
The new scanner (flex based) only has a single context, the new parser is hand
written instead of bison based.

The old bindary was 966 960 Bytes, the new one is about 6% of this 59 728 .
The new scanner code is a lot easier to read, the new parser code is
less lines of C than it had lines of bison grammer, it is earsier to read
as well.

The new code executes about 10 times faster, than the old one.

The new code also has the extensive error reporting capability, that
drove the old implementation into its bloatness.