r2437: Ernest Montrose (stratus):
This patch fixes an issue where testsuite.pl hangs and timeouts waiting
for result from a drbd device under test.  When this happens both the
agent and the client are waiting for data from each other.
Testsuits.pl eventually fails with a timeout.  Send_command() changed to
synchronize the I/o and close the descriptors when done.