drbdadm: don't try to set res->peer too early; rename --to to --peer
set_peer_in_resource may set config_valid=0 and exit.  bailing out from
post_parse() because of unrelated (to the current invocation) resources
is bad.

try to set_peer_in_resource as late as possible.

also rename --to to --peer
(it got renamed from --host to --to earlier,
but that renaming was incomplete anyways)