drbd.spec.in: make the "drbd" package depend on all the spun-off packages
Now that we have split DRBD's userland and integration facilities into
several packages, define the "drbd" package as a virtual package that
depends on all others (except drbd-km; that exception was true also
for the old "drbd" package). Thus, when users upgrade the "drbd"
package from older versions, all the new packages will be pulled in.

We need to have at least one file in a %files section for the "drbd"
package, otherwise it won't be built at all.