drbdadm: make references to non-existing resources non-fatal
When deleting resources, it is an easily made
oversight to leave references to the deleted
resources in sync-after statements of other resources.

Don't fail on every pacemaker-induced action, as it would
trigger monitor failure, stop failure and
ultimately lead to stonith.

If some resource references a non-existing resource
from a sync { after .*; } statement,
this is still a "config invalid", if run from a TTY.

If we don't have a controlling TTY,
i.e. we are run from a script,
only complain to stderr, but keep running.