drbdadm adjust cosmetic: do nothing, if there is nothing to do
To cater for the corner case of someone renaming a
resource definition without changing anything else
in the config file, we send a no-op syncer command
to the kernel, so the kernel will send a udev change event,
and udev can update the /dev/drbd/by-res/* symlinks.

Unfortunately we sent out this no-op syncer command
even if really _nothing_ had to be done.

Impact: cosmetic ;)