user: Introduce options with an explicit yes/no value in config files
If options like no-disk-barrier were used in the common section, there
was no way to turn those options on again in an individual resource.  To
fix this, introduce additional keywords which take an optional yes | no
argument.

For example, "no-disk-barrier" is equivalent to "disk-barrier no", and
is the opposite of "disk-barrier yes" and "disk-barrier" without
argument.