drbdadm: Fix resource and volume registration
When a new connection or a new volume is created ("drbdsetup <name>
new-connection", "drbdsetup <minor> new-minor"), we remember where to
find the config file used to help the callbacks from the kernel into
user space.  This was so far done in drbdadm.  Move un-registering from
there into drbdsetup to avoid stale entries, and register connections /
volumes when creating them instead of checking/updating the database
with almost every command.

No longer check if drbdadm is called with a different config file than
the one used when creating a resource / volume: this check doesn't
really help; rely on the user not to do do stupid things instead.