drbdadm: in adjust, fix _is_default detection for new yes/no flags
"ro" was yet an other badly chosen variable name.

Rename to "current_option", as that is where it points to.
As it is used to set the current_option->is_default member,
it better points to the currently parsed and freshly appended option,
so we need to do the current_option = new_opt() explicitly.