drbdadm: before blocking the boot process, try to put /dev/console into sane mode
When some boot splash thingy redirects the system console,
we have been out of luck when waiting for a "yes\n" confirmation
on our prompt " To abort waiting, enter yes ...".

We now try to disable plymouth and usplash, and put the console into
"canonical" mode, before blocking on that prompt.

In case our tty is stuck in "raw" mode despite of that,
we will typically get characters one at a time,
so we try to cope with input in that mode, too.