drbdsetup new-minor: Fix argument passing to the kernel
The new-minor command takes a resource name as its first, and a minor number as
its second argument.  To allow this calling convention without ugly hacks, pass
the drbd_genlmsghdr to the argument convert functions; this is where the minor
number is passed to the kernel.

Make the resource name and minor number regular arguments of the new-minor
command: otherwise we would end up with multiple drbd_cfg_context  structs on
netlink, which the kernel can't handle.