build: include the kernel-version in the kmod package version on rhel as well
RHEL 6.1 dumped BARRIER, and backported FLUSH/FUA.
They did this in a kABI "compatible" manner, but
using a module compiled against the "wrong" headers
obviously cannot use either feature, and triggers kernel warnings.

To avoid splitting up the repository, we need to be able to include
several versions of the kmod-drbd package.