drdbadm: Fix drbdsetup option pass through with "drbdadm -- <drbdsetup options>"
This broke when switching to processing non-options in line with options in the
getopt_long() loop: interestingly, getopt_long() does not process options after
"--" in line in that case.  Stop processing non-options in line: we no longer
need to.

Also, improve the error messages that drbdadm spits out, and return a status of
0 for "drbdadm help" and "drbdadm help <command>".