packaging: No longer try to be clever with the drbd-rgmanager dependencies
Dependencies for drbd-rgmanager are particularly awful. On RHEL 5
and prior (and corresponding Fedora releases), %{_datadir}/cluster
was owned by rgmanager version 2, so we have to depend on that.

With Red Hat Cluster 3.0.1 (around Fedora 12), the DRBD resource
agent was merged in, and it became part of the resource-agents 3
package (which of course is different from resource-agents on all
other platforms -- go figure). So for resource-agents >= 3, we must
generally conflict.

Then for RHEL 6, Red Hat in all their glory decided to keep the
packaging scheme, but kicked DRBD out of the resource-agents
package. Thus, for RHEL 6 specifically, we must not conflict with
resource-agents >=3, but instead require it.

The saga continues:
In RHEL 6.1 they have listed the drbd resource agent as valid agent,
but do not include it in their resource-agents package. -> So we
drop any dependency regarding rgmanager's version.