drbd: Clarify which enum drbd_conns values refer to what
Keep values which refer to the connection state (connection->cstate) prefixed
with C_ and prefix values which refer to the replication state with L_.

The previous value C_WF_REPORT_PARAMS is used to indicate that a network
connection is established and that the device parameters are to be exchanged.
Replace this with C_CONNECTED for the connection state, and L_STANDALONE for
the peer device state, both with the same value.

The previous C_CONNECTED state becomes L_CONNECTED.