drbd: Move max_buffers setting from net_conf to device_conf
This setting is defined by the "drbdsetup new-minor" command; there currently
is no command to change it for an existing minor.

TODO: Update the parser and documentation.  Check how the setting is used in
the code.  This setting limits the number of pages (independent of PAGE_SIZE)
plus trhe number of requests that can be active.