drbdmeta: fix wipe-after-convert to only wipe if necessary
With 07, we used to have a "fixed" meta data layout,
where our "super block" was located at (end-of-device - 128M).

We then introduced the "flexible" meta data layout,
where our "super block" is located at (end-of-device - 4k).

Whenever we convert from "fixed" to "flexible" layout or back, we want
to wipe the previous super block location, just to be sure it won't be
accidentally detected and used.

This logic broke when introducing the conversions to v09,
where it now could happen that when converting from v08 meta data,
it would first write a correct v09 superblock, but then wipe it out.