drbdadm sh-udev: Prevent null-pointer access on diskless nodes
On diskless nodes, no lower-level device is configured.  In this case, don't
try to create a /dev/drbd/by-disk/... symlink.  This requires to pass the list
of symlinks to udev instead of constructing that list in scripts/drbd.rules.

Also, fix what the documentation says about /dev/drbd/by-disk/ symlinks.