drbdmeta: default to open O_EXCL
For internal (and "flexible" external) meta data,
default to open the meta data device O_EXCL, to recognize in-use
devices, even if the minor number given on the command line is
accidentally wrong, or the device is in use by something other than drbd.

You can still override it with --force,
e.g. to peek at the data with dump-md.

For old-style fixed size indexed meta data, we cannot use O_EXCL,
as in general it will be partially in-use by attached DRBD minors,
but we still want to manipulate the slots of the non-attached minors.