crm-fence-peer.sh: implement --network-hickup timeout
If cib and crmadmin -S think the peer is still reachable,
optionally wait for this timeout, before looking at the DRBD state.

Maybe it was just a glitch, and we have been able to recover
without the need for resource or node level fencing.

This possibly avoids spurious node level fencing events
in some (dual-primary) setups.