crm-fence-peer.sh: guess whether or not pacemaker will fence
If there won't be a fencing operation,
things will not improve by waiting for the full timeout.

So once the peer is known to be unreachable,
and there won't be any fencing,
stop polling the cib.

If we detect that a node has already been fenced
(or cleanly shut down), also report "fenced" instead of "unreachable".