drbd: rpm compat: fix /sbin/drbd* -> /usr/sbin/drbd* symlinks
When upgrading from previous tool versions, the symlinks got created in
the postinst stage of the new package, only to be removed immediately
after that due to "ownership" of these files by only the old filelist.

Add explicit symlinks to the new filelist again,
drop the postinst scriptlets, and make it an rpm conditional.