No longer do a DNS name resultion on the hostname
We have to assume that there are too many systems out there that
can not resolve their own hostname.

As the hostnames_equal() function takes into account that one name
might have a domain, and the other might not have a domain
caninization of the hostname has no point at all.