LLVM scan-build: Dead assignment
drbdadm_main: the variable "rv" is loop local and setting it before
calling continue does not make sense. Additionally, the return value of
a write to fileno(stderr) is not checked. This should be fine and we
simply void-cast the return value. Further, the return value of
waitpid() is not evaluated and now gets casted to void. This should at
least give others a hint.

drbdadm_usage_cnt: return value of fgets() is never evaluated. In this
case it is just a "press any key" message, so we do not care.

drbdmeta: return value of lk_bdev_load() is never evaluated. We handle
this case as other fauls in this function and return -1.

drbdsetup: there was a mix of "err" and "rv". "err" got set, but was
then reset at the end with a call to check_err(rv, desc). We removed
"rv" and finally set "err = check_err(err, desc)".

drbdtool_common: The variable "skipped" gets reset if it was set. As the
variable is not static, it does not make sense.