LLVM scan-build: Value is garbage/undefined or Uninitialized argument
drbdadm_main: Create the pipe() in any case. This helps the analyzer
because it has no special knowledge which combination of arguments are
allowed and therefore derives that "fd" might be set to a "pipe_fds[0]"
and therefore to a pipe end which was never created.

drbdmeta: The variable "lw" is only set if a condition is true, and only
printed if the same condition is false. But on the first iteration the
condition is true and therefore the value is set for all iterations that
follow. Never the less, the analyzer does not know this, so initially
set the variable to 0.