LLVM scan-build: Dereference of null pointer
drbdadm_main: variable "s" could be a NULL pointer and it will get
dereferenced by accessing "s[rr]". Simply check if "s" is not NULL.

drbdsetup: Frist it is checked if "cmd" is not NULL. If it is NULL a
function that exits is called. If it is not NULL "cmd" gets
dereferenced. The analyzer does not know that the function that gets
called before exits in any case. Therefore give the analyzer the
necessary hint.