LLVM scan-build: Memory leak
drbdmeta: "all_cpu" is function local, if we return before setting
values from it to "hot_extent", it is safe to free the memory of
"all_cpu".

drbdsetup: "iov.iov_base" is malloced, therefore it is NULL or has valid
memory. In both cases it is save to free() it. The same applies to
"smsg". msg_free() also handles the NULL case. So if one is not
successfully malloced free them both. "key" is NULL or malloced, so we
can also free it in any case. free() handles the NULL case.