LLVM scan-build: Dead assignment
libgenl: "g" is always set, but only printed if the dbg() macro
evaluates to true. With the current value it evaluates to false and
therefore setting "g" leads to a dead assignment. By introducing a bit
of preprocessor magic we can avoid that the code is generated. This
patch is not particularly beautiful, but the scope is small enough.