LLVM scan-build: Allocator sizeof operand mismatch
drbdadm_main: resource_names is of type "char ***", so we increase
"*resource_names" by allocating new elements with sizeof(char*), not
with sizeof(char**). While at it Philipp Marek spotted that the loop is
overly complex anyways. We do not need to play this realloc game, we
have a save bound for the number of resources given on the command line,
namely "argc".