drbd: Fix regression in init script, partial revert of f064a3d4b676a
While drbd9 supports automatic promote and "sh-b-pri" was removed from
drbdadm, we should keep the call to "sh-b-pri" in the init script if it
is used with drbd8. In case of drbd9 the command will fail (because it
is no longer supported), but we ignore the return value.

Thanks to Apollon Oikonomopoulos <apoikos@debian.org> reporting it.