Fix node_id auto auto-generation for config with proxy
If the proxy is running on the same host it is very likely
that the same local address is used on both machines. E.g.

resource r {
	device minor 0;
	disk /dev/vdb;
	meta-disk internal;

	on rckdeba {
	        address ipv4 127.0.0.1:7000;
	        proxy on rckdeba {
			inside  ipv4 127.0.0.2:7000;
			outside ipv4 10.43.70.115:7000;
		}
	}

	on rckdebb {
		address 127.0.0.1:7000;
		proxy on rckdebb {
			inside 127.0.0.2:7000;
			outside 10.43.70.116:7000;
		}
	}
}

Up to now this produced the internal error:
"BAD LUCK, equal hashes".

If that happens, and we have a proxy section on both machines, use
the proxy outside addresses instead.