Fix adjust (v9) to not crash when a proxy is involved
The obvious bug was that command that expect a connection
to be set in the ctx arguemnt where called without one set.

But right now every connection leaving a node with a
proxy definition will go through the proxy. It is obvious
that we need to change the configuration syntax in that regard.
With that the data-structure in drbdadm ... phew