drbdadm: Added support for multiple connections
This commit adds support for 'drbdadm connect' if called with multiple
connections. Previously, it was sufficient to find a matching connection
and set the ctx. With multiple connections, we have to do more work,
e.g., we have to check if a given subset of connections would enable
connections that have been already enabled by another subset of given
connections.

A hopefully elegant way to setup connections is to filter them and set
the ignore flags accordingly, and finally bring up those connections
that are not ignored. Our codebase was ready for that. Before we can set
the ignore flags, we have to check certain conditions: Would we enable
an otherwise disabled connection? Would we enable a connection multiple
times? Setting/and resetting the original "conn->ignore" flag is error
prone. Checks should not touch the original ignore flag. Therefore,
introduce a conn->ignore_tmp flag that can be manipulated by checks, but
is otherwise ignored by the rest of the code.

Currently, only ctx_by_name makes use of these checks to check
connections, in the future the check-states can be used for
resources, volumes, and so on. Basically, for everything where we have
to check if we would enable something disabled or if we would enable
something multiple times.

The following states have been introduced (note: although I use the word
'connection', it should not be limited to connections in the future):

WOULD_ENABLE_DISABLED: would the command enable an otherwise disabled
connection?
WOULD_ENABLE_MULTI_TIMES: would a connection be enabled multiple times?
SETUP_MULTI: do the filtering and set conn->ignore accordingly. This is
usually the final state, if all checks succeeded.
CTX_FIRST: set ctx on first valid connection. This is the way things
worked, code paths that need a ctx can use this semantic.

Tested with:
$ ./user/drbdadm connect -c ./mulcon.res -d r0
$ ./user/drbdadm connect -c ./mulcon.res -d r0:bob
^^ also includes the unnamed connection
$ ./user/drbdadm connect -c ./mulcon.res -d r0:charlie
$ ./user/drbdadm connect -c ./mulcon.res -d r0 r0:charlie
^^ ERR: because would enable a connection multiple times
$ ./user/drbdadm connect -c ./mulcon.res -d r0:ab0
$ ./user/drbdadm connect -c ./mulcon.res -d r0:ab1
$ ./user/drbdadm connect -c ./mulcon.res -d r0:ac0
$ ./user/drbdadm connect -c ./mulcon.res -d r0:ac1
$ ./user/drbdadm connect -c ./mulcon.res -d r0:sametwo
^^ brings up both connections (to bob and to charlie)
$ ./user/drbdadm connect -c ./mulcon.res -d r0:three
^^ ERR: would match a connection between bob and charlie
^^ we could filter only valid conns, but strictly speaking,
^^ this is not what the user asked for.
$ ./user/drbdadm connect -c ./mulcon.res -d r0:notexist
^^ ERR: because not a valid connection
$ ./user/drbdadm connect -c ./mulcon.res -d r0:ab0 r0:ab1
$ ./user/drbdadm connect -c ./mulcon.res -d r0:ab0 r0:ab0
^^ ERR: because would enable a connection multiple times
$ ./user/drbdadm connect -c ./mulcon.res -d r0 r0:ab0
^^ ERR: because would enable a connection multiple times
$ ./user/drbdadm connect -c ./mulcon.res -d r0:ab0 r0
^^ ERR: because would enable a connection multiple times
$ ./user/drbdadm connect -c ./mulcon.res -d r0:bc0
^^ ERR: because it is a disabled connection (not part of bc0)

Currently, there is a second code path that uses ctx_by_name, if you
uncomment the commented lines, it still fails and tells us that:
'resync-after contains 'r1', which is ambiguous, since it contains 2
volumes'

resource r0 {
	volume 0 {
		device    /dev/drbd1;
		disk      /dev/sda7;
		meta-disk internal;
 #		disk {
 #			resync-after r1;
 #		}
	}

	on aperol {
		node-id   0;
		address   10.43.70.1:7000;
	}
	on bob {
		node-id   1;
		address   10.43.70.2:7000;
	}
	on charlie {
		node-id   2;
		address   10.43.70.3:7000;
	}

	connection ab0 {
		host      aperol  port 7000;
		host      bob    port 7000;
		net {
			protocol C;
		}
	}
	connection ab1 {
		host      aperol  port 7001;
		host      bob    port 7001;
		net {
			protocol C;
		}
	}
	connection {
		host      aperol  port 7002;
		host      bob    port 7002;
		net {
			protocol C;
		}
	}
	connection {
		host      aperol  port 7003;
		host      bob    port 7003;
		net {
			protocol C;
		}
	}
	connection bc0 {
		host      bob  port 7004;
		host      charlie    port 7004;
		net {
			protocol C;
		}
	}
	connection ac0 {
		host      aperol  port 7005;
		host      charlie    port 7005;
		net {
			protocol C;
		}
	}
	connection ac1 {
		host      aperol  port 7006;
		host      charlie    port 7006;
		net {
			protocol C;
		}
	}
	connection samethree {
		host      aperol  port 7007;
		host      bob    port 7007;
		net {
			protocol C;
		}
	}
	connection samethree {
		host      aperol  port 7008;
		host      charlie    port 7008;
		net {
			protocol C;
		}
	}
	connection samethree {
		host      bob  port 7009;
		host      charlie    port 7009;
		net {
			protocol C;
		}
	}
	connection sametwo {
		host      aperol  port 7010;
		host      bob    port 7010;
		net {
			protocol C;
		}
	}
	connection sametwo {
		host      aperol  port 7011;
		host      charlie    port 7011;
		net {
			protocol C;
		}
	}
}

 #resource r1 {
 #	on aperol {
 #		node-id   0;
 #		address   10.43.70.1:7012;
 #	}
 #	on bob {
 #		node-id   1;
 #		address   10.43.70.2:7012;
 #	}
 #	volume 0 {
 #		device    /dev/drbd1;
 #		disk      /dev/sda8;
 #		meta-disk internal;
 #	}
 #
 #	volume 1 {
 #		device    /dev/drbd2;
 #		disk      /dev/sda9;
 #		meta-disk internal;
 #	}
 #}