rhcs_fence: Do not invoke fence_node multiple times
fence_node, or more precisely the fence agents, are
very likely to fail in ugly ways if they are invoked
in parallel in multiple instances.

The 'cman_tool kill -n $remote_node' command causes
fenced to fence the remote node anyways. In case fenced
is doing that, simply wait for that to finish...