crm-fence-peer.sh: logger: avoid subshell
Bash, or some versions of it, may decide to involve a subshell
for the output redirection to logger, so the process tree would look
like
  main script
    sub-shell
      logger

Sometimes, the sub-shell would linger around,
and keep logger from exiting, even if the main-shell finished already.

The second exec replaces the subshell, and logger will notice directly
when its stdin is closed once the main script exits.
This avoids the spurious logger processes.