drbdsetup-84: fix events command (would always abort)
The events command has a ctx_key of CTX_MINOR|CTX_RESOURCE|CTX_ALL.
Which is now missing from the new switch statement in generic_get(),
and ends up in the default: assert(0) :-(

Also, the detection of minor vs resource in main() was broken, and
would assign CTX_MINOR, even if minor would still be -1U (unknown).

Correctly assign CTX_RESOURCE if the object name does not correctly
resolve to a minor number.

Since that context variable is global already, the switch statement
then can deal with this known context information, and does not need to
guess around based on the allowed context variants for the command.