drbdsetup events2: Improve how timestamps are assigned
The --timestamps option of command events2 did assign each notification
a timestamp.  Improve this so that notifications which refer to the same
event (for example, an atomic device and peer device state change) get
the same timestamp.

In the initial state report, all notifications referring to the same
resource are considered one event and get the same timestamp; separate
resources and the terminating "-" notification get different timestamps.