drbd init script: on start, first try to load the module
On older systems, if the module is not loaded yet,
and in your current directory happens to be a directory or file named
"drbd", modinfo -F version drbd may error out.
In that case the automatic drbd driver version detection fails,
and we may get a parse error for certain special configuration tokens
present in 8.3 or 8.4, but no longer in "future" drbd config files.

If we first load the module, the version detection uses
the content of /proc/drbd instead.