exit codes: redefine E_USAGE to 1 (not 3)
Avoid an exit code of 3.
In case this was a fence-peer handler invocation, this would be
interpreted as "fencing successful/not needed, peer has bad data anyways".

If for some stupid reason the automatic version detection does not work,
this could actually happen, e.g. if a drbd 8.3 module calls
"drbdadm fence-peer minor-0", but the more recent drbdadm would expect
a resource name, not a minor number, and exit E_USAGE.