drbdadm/meta/usage_cnt: Usability bug where output is not shown
If a user executes "drbdadm create-md r0" and the device backing r0 is
currently mounted, drbdadm seems to hang. "drbdmeta" is called twice:
Once with "read-dev-uuid" and once to actually write the data. The first
call is meant to be invisible (besides reading back stdout), therefore
"run_admm_generic" supresses stderr. Opening the mounted device fails
and "drbdmeta" asks the user for confirmation to open it anyways. As
stdout is not printed, the user does not see the question and for him it
looks like "drbdmeta" hangs.

This patch does not suppress stderr (why should we even want that?) and
prints the question to stderr. This keeps the current logic intact, that
"read-dev-uuid" prints its output to stdout and "adm_create_md" reads it
back.

This is a quick fix, but we should rethink this whole "interactivity"
scheme.