drbd main ipv6: Fix interface indices larger than 255
fscanf line probably copied from kernel printf, but printf specifies the
minimal field width and scanf the maximal field width.

The bug was triggered on Xen servers that create and assign new,
incremented interface IDs for new machines.

ifindex is of type int anyways: struct net_device as defined in
include/linux/netdevice.h. We should not rely on the assumption that
the values are small enough to fit in one byte.

Signed-off-by: Roland Kammerer <roland.kammerer@linbit.com>