drbdsetup-84: fixup status subcommand
The recently [1] introduced (actually backported from DRBD 9)
"status" sub command needed some minor fixups when interpreting and
reporting connection states.

The connection state enum symbols have been slightly renamed in DRBD 9,
which means to have the same meaning, we need to use different names in 8.

We also drop the "IP pair" information from the output,
because we are about to change that in DRBD 9 to the "peer node id".
In DRBD 8.4, we will only print "peer" instead.

[1] post drbd 8.4.5, post drbd-utils 8.9.1, not yet released