systemd unit: install sysv init script to /lib/drbd/drbd, and call that
You are not supposed to use the init script or systemd unit file anyways.
Please use a cluster manager (Pacemaker) instead.

If for some (hopefully good) reason you insist to have your init system
"start" and "stop" DRBD, there is not much point trying to re-implement
all the logic the init script developed over the years in some systemd
unit file.

Call the init script from the unit file instead.

Why not just keep using the init script,
and the SysV-compat mode of systemd?

Because we need the network to be up and configured already,
and some network scripts are still SysV init scripts as well,
but return "success" even if the network is not yet fully up.
Which would then cause our script to fail.

Also require both network and sshd to be up, so an admin will be able to
log in to kill "endless waiting" drbdadm processes waiting for the
initial connection(s) (which is a feature of the init script).