drbdadm wait-con-int: fix busy loop in case we cannot find a tty
If for some reason (hi there, selinux) we don't have a tty, and cannot
open /dev/console, or the tty we had closes on us, don't go into a busy
loop trying to read from an effectively closed file.

Also use the same wording for the "banner" message for both v84 and v9.