drbdsetup: sh-status: recognize a data-less primary as configured
Before, a diskless, disconnected (data-less) primary would be reported
as unconfigured.

This has implications for the ocf drbd resource agent:
in this situation, it would before (incorrectly) report "not running",
now it (correctly) would report "running master",
but still remove the master score.

Outcome is effectively the same:
Pacemaker is supposed to run recovery actions on this resource.