Do not try to configure peer_device_options implicitly
We do not know if all the connections are already established
when creating a new minor. Neither do we know if all the minors
are there when creating a connections.

The idea was wrong.

Keep the new-minor and the connect commands clean and small.

Have the up command to schedule the peer_device_options at the right
time.