drbdadm: read configuration files only once.
This makes it possible to have 'include "global_common.conf"' in
each resource file, and to use only these configuration files
with the resource agent.

This way a problem with other configuration files (eg. a syntax error
during creation of a new resource) won't cause the RA and Pacemaker
to barf (and switch over).