drbd usage count: don't re-register each time the module is unloaded
If the module currently is not loaded,
we guess the module version from "modinfo -F version drbd",
and have no means of determining the "git_hash".

If we then re-register just because the "git_hash" (all zeros) is different
from the non-zero git hash we stored last time the module was loaded,
this becomes very annoying.

Skip re-register, if we currently don't know the module version control hash.