drbd usage count, v83/v84: don't re-register each time the module is unloaded
If the module currently is not loaded,
drbdadm 8.3 and 8.4 fall back to their own version git-hash.
Which now, since the kernel/utils split, is different from the module hash.

If we then re-register just because the "git_hash" flapped to the "utils" hash
from the hash we stored last time the module was loaded, or vice versa,
this becomes very annoying.

Skip re-register, if we currently don't know the *module* version control hash.