debian: Fixed systemd related packaging bug
On a clean Debian/Jessie package building failed because:
- dh_systemd is not installed by default, not even if "devscripts" are
installed. Therefore, the debian/rules file guesses it is not a system
managed by systemd. Therefore, ./configure is executed without a
initsystem hint. This then leads to a situation where ./configure tries
to guess the init system. As /sbin/init points to systemd, it guesses
systemd.

=> configure thinks its on a systemd system vs. debian/rules, which
thinks it is on a non-systemd system. The packaging system then tries to
install a file that does not exist, because the autotools did not
generate it, because they thought they are on a systemd system and do not
need to build sysv init style files.

The obvious fix is to add dh-systemd to the "Build-Depends". This does
not help in our specific situation, as we want to support old, but still
supported Ubuntu systems. Therefore we introduce a specific control file
for Ubuntu/Precise, which is the only supported Ubuntu release that does
not have a dh-systemd package.

tl;tr: If you are on a Ubuntu/Precise system, you have to move
debian/control.ubuntu-precise to debian/control before building the
package.

control.ubuntu-precise can be deleted 2017-04 ;-).

If this ever makes problems we revert this commit and use the following
strategy:
- We/LINBIT knows what to do, we install dh-systemd in the change-roots
where we want systemd support.
- Users that use a git checkout have to take care by themselves if they
want/need dh-systemd.