dump: correctly dquote escape strings the scanner won't tolerate
The intention of dump is to produce a normalized, valid config file.
Dumping the output of dump again should produce the same output.

For certain values of "shared secret" (and other string values),
this was not always the case; dump would strip the double-quotes
of the original, resulting in parse errors when using the dump
as config file.

The STRING token in the scanner is defined as [a-zA-Z0-9/._-]{1,80}.

For drbdadm dump, that means
if (0 == fnmatch("*[!a-zA-Z0-9/._-]*", str, 0) || strlen(str) > 80)
	double quote;