v84: new disk config option rs-discard-granularity
When rs-discard-granularity is set to a non zero, positive value then DRBD
tries to do a resync operation in requests of this size.  In case such a
block contains only zero bytes on the sync source node, the sync target node
will issue a discard/trim/unmap command for the area.
The value is constrained by the discard granularity of the backing block
device. In case rs-discard-granularity is not a multiplier of the discard
granularity of the backing block device DRBD rounds it up.  he feature only
gets active if the backing block device reads back zeroes after a discard
command.
The default value of is 0 (disabled).
This option is available since 8.4.7.