doc: Makefile update to avoid infinit loop in parallel build
For reasons not completely understood, on the second "make -j",
make would more often than not end up in an infinit loop inside the
documentation/v84 or /v9 directories.

Rework the Makefiles there to be more explicit about the dependencies.

Be explicit about the files for html or pdf as well.

In v84, add status and events2 to the to-be-processesed command list.

Also add explicit no-op rules for files tracked in git
to avoid useless scanning for implicit rules to rebuild them.